Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21127 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/opengl/gl_polyrender.cpp
@@ -44,7 +44,7 @@ csGLPolygonRenderer::csGLPolygonRenderer (csGLGraphics3D* parent)
 {
   SCF_CONSTRUCT_IBASE(0);
   csGLPolygonRenderer::parent = parent;
-  renderBufferNum = ~0;
+  renderBufferNum = (uint)~0;
   polysNum = 0;
   buffer_accessor.AttachNew (new BufferAccessor(this));
   shadermanager = parent->shadermgr;
